+++ R/05-model.elements.R
@@ -137,11 +137,13 @@ model.elements <- function(model, factorNames, itemtype, nfactNames, nfact, J, K
                     guess=guess, upper=upper, fulldata=fulldata, J=J, K=K, customItemsData=customItemsData,
                     nfact=nfact+length(prodlist), parprior=parprior, monopoly.k=monopoly.k,
                     parnumber=parnumber, estLambdas=estlam, BFACTOR=BFACTOR, item.Q=item.Q,
-                    mixed.design=mixed.design, customItems=customItems, key=key,
+                    mixed.design=mixed.design, customItems=customItems, key=key, factorNames=factorNames,
                     gpcm_mats=gpcm_mats, spline_args=spline_args, itemnames=itemnames)
+
     ret[[length(ret) + 1L]] <- LoadGroupPars(gmeans=gmeans, gcov=gcov, estgmeans=estgmeans,
                                              estgcov=estgcov, parnumber=attr(ret, 'parnumber'),
-                                             parprior=parprior, Rasch=all(itemtype %in% c('Rasch', 'rsm', 'Tutz')),
+                                             parprior=parprior,
+                                             Rasch=all(itemtype %in% c('Rasch', 'rsm', 'Tutz', 'PC1PL')),
                                              customGroup=customGroup, dcIRT_nphi=dcIRT_nphi, dentype=dentype)
     attr(ret, 'prodlist') <- prodlist
     attr(ret, 'exploratory') <- exploratory